+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ Change Log
 =====
 * Support for viewing and editing user projects and institutions (C4-945).
 * Support for viewing info about VPCs, security groups, subnets, et cetera (C4-961).
-* Support for viewing info about stacks.
+* Support for viewing info (outputs, parameters, resources, templates) about stacks.
 * Support for viewing checks by schedule.
 * Support for simple checks search.
 * Support for simple users search.